Fix sliceIndex implementation for Slice (Any sh)
The implementation of convertSliceIndex uses unsafeCoerce.
While testing the use of the repN function (defined in
SliceExamples.hs) I found a case where the type equality didn't
hold.

   Repr.SliceShape (EltRepr (Any sh))
:: Repr.SliceShape ()
:: ()

but

   EltRepr (SliceShape (Any sh))
:: EltRepr sh

These are clearly not the same in general.

I created a new class ShapeElt which represents the types
of values that can be shapes (not slice specifiers). I added
two new type families. The first,'SliceAnyRepr' maps from types, 'sh',
satisfying 'ShapeElt' to the representation of the slice
'Any sh'.

e.g. Any (Z:.Int) is effectively the same as Z:.All which
is represented by ((),()) (in module D.A.A.Array.Representation)
Therefore 'SliceAnyRepr (Z:.Int) :: ((),())'

The other type family is ShapeEltRepr which was added
for future purposes because I've got an idea about how to
remove the need for 'convertSliceIndex' and place
more restrictions on what can be shapes and slices in
constructors such as Replicate.

With this fix function 'repN' now produces correct results.