fix segmented scan family
The key missing feature here was support for zero-length segments. As soon as one of these was encountered, the rest of the data would be thrown out of whack. To counter this, the flag arrays that mark the start (or end) of each segment can now contain not just zero or one, but non-unity values that mark the start (end) of several empty segments. The remainder is fairly standard, although still required a rewrite.

fixes AccelerateHS/accelerate#44