cleanup, wibbles
Clean up the code a bit. Also try to improve performance a bit by forcing the first step (findClosestCluster) to be evaluated, instead of fusing it into its consumer, which results in work duplication. Unfortunately, this is hitting a sharing recovery bug in Accelerate, so can't be used right now.