package the common test-harness modules into a library
This avoids recompiling the modules for every test program, and means that dependency changes need only happen once.

This also handles the common command line arguments, such as choosing the backend or displaying the help message. The library also becomes the only component to directly import the individual backend implementations.