avoid unused import warnings with this one weird trick!
As the unused import check is done top-to-bottom, explicitly importing the Prelude at the bottom of the import list is usually enough to suppress the unused imports warning on 7.10, where extra stuff found in e.g. Data.Functor and Control.Applicative in earlier versions of ghc, is now available from Prelude.