Use an environment when pretty printing expressions
Instead of just taking the current on input the current de Bruijn level as an int, require a valuation environment in the usual style that carries the bound variable name.

It is possible to create the environment as necessary, but because our environment type is implicitly nested tuples rather than fixed to, say, type level lists, we require a class constraint to be able to do that. This means I had to disable debug printing at a few places, which is unfortunate.

The motivation is that I want to be able to reuse the pretty printing infrastructure while supplying my own names.

One further change in this patch is that I have disabled some bits of pretty printing of sequence expressions. It didn't look like those were done very well, so I want to rethink that.