First attempt at Graphviz output of Accelerate programs
There are some configuration options to show the entire scalar expression, or just the type of the node, as well as to draw lines edges between nodes for every array access or just the unique dependencies.

This is currently limited to DelayedOpenAcc (rather than generalised to PreOpenAcc) and does not yet cover all of the Acc node types (Apply, Awhile, Acond).