Don't force let binding manifest array arguments to permute, stencil
In particular, if the defaults array to a forward permutation was not already let-bound, then this was the only use of that array and we can update the results in-place, avoiding an additional memcpy.

Updates & improves the graphviz module as appropriate.

Reverts: 248222b66ba7c33e5b47756b486e29afff1cfe1f
Unifies: a93028b6fba78f0191477939fc11fa4c24673868