Rename infix operators to have the standard names
Previously we redefined operations such as (<) as (<*), as their types changed and thus we couldn't provide an instance for the standard Prelude.Eq class.

However, since we now redefine the standard class hierarchy, rather than using (<*) at a fixed (closed) type (IsScalar), there is no longer a good reason to do this.

This change will break all existing code, but so did the introduction of the class hierarchy, so we might as well use the same names as in standard Haskell.