mandel: cache the image
We only update the image on user interaction, so it is worthwhile caching it.

However, this is mostly this is to get around a nasty graphical bug when using the PTX backend. It seems like gloss updates the image correctly, but then on the next frame it could just be garbled (just rubbish, uninitialised data). So maybe gloss is copying the data twice when it shouldn't be, the second time being after the array has been deallocated? Not sure...