move stencilAccess function
The implementation and placement of the stencilAccess function was unsatisfactory, and not reusable for any other backends (it is really only there to make implementation of the interpreter easier). Move it to the interpreter module and implement in terms of the StencilR GADT, as the real backends need to do.