make the AST show instances a bit smarter
If we assume that we are show-ing an Acc or Exp expression to the terminal (which seems the most likely use case?) then we can be a bit cleverer in how the expression is rendered. In particular, try to pick up the width of the terminal as well as whether or not it supports ANSI colourisation.