Allow stencil fusion
Right now, this just duplicates the work of the source (delayed) array at each access to the stencil pattern. In future we should add extra logic to determine when that should be done and when not; for now we leave that to the user (since they currently have the ability to prevent fusion, but not force it).