leave terms at their lowest common use site
The motivation is that we can determine when it is safe to in-place update arrays as a straightforward syntactic check. The disadvantage is that this removes the invariant that the arguments to terms have been evaluated already (that is, applied as array variables).

SEE: [Fusion and the lowest common use site]

AccelerateHS/accelerate#432