drop the convert-segment-offset pass
The internal representation of FoldSeg now expects the segment descriptor to be in offset format, rather than encoding the length of each logical sub-array. The existing `foldSeg` function now unconditionally inserts the `scan` of the segment descriptor. A new `foldSeg’` takes the offset segment descriptor directly.

We lose the ability to execute a sequential foldSeg without scan. We could reintroduce this by keeping track on a per-FoldSeg node basis what format the segment descriptor is in.