Pretty: Add parens around Let more often
Previously, the parentheses requirements of Let and Alet were written as
if they were function applications. However, this is untrue, since the
body expression of a let-binding extends to the right as far as
possible.

This fixes the printer to _always_ add parentheses if there is a
non-trivial expression context around the let-binding. We could be more
conservative by allowing a context on the left only, but I thought this
would be less surprising / more consistent.