Pretty: Align T<n> arguments
This is not _necessary_ so much for prettyTuple and prettyAtuple, but
it's very much wanted for prettyLhs, because it makes the code closer to
parseable Haskell. Without the 'align' there, the arguments to the T<n>
sometimes came on the same indentation level as the constructor, meaning
they're seen as declarations in the Let due to layout rules -- which
doesn't parse.