Add annotation verbosity option to pretty config
With this we should be able to hide most of the verbosity when it's not
needed, and show more information when it is. Because the option is
needed everywhere where annotations are printed, this does mean that we
get to pass around the PrettyConfig to basically every pretty printing
function.