Remove use of stable names from graphviz algorithm
The stable names were not actually used non-trivially, and resulted in
incorrect graphs with duplicate node ids and confused edge targets due
to (I think) the assumption that hashStableName always generates unique
names. Since the hashes seem to always carry only one byte of
information, that's bound to go wrong.

Symptoms of this issue: the printed graph has self-loops or back-edges,
making it not a DAG anymore. Sometimes it is very clear in the .dot file
because there are multiple different node declarations with the same id.

I analysed all the use of stable names (through mkNodeId (in this commit
renamed genNodeId)), and found that all usages are on objects of type:
- DelayedOpenAcc, the Manifest constructor. This is done at most once
  for each node in the AST.
- PreOpenAcc, only the Apair constructor. This is done at most once for
  each Apair in the AST.
- ALeftHandSide (when handling an Afun). At most once for each Afun.
- PNode, once on the result node of the whole program, and (in replant)
  once for the output graph node of AST nodes in certain positions (i.e.
  where replant is called).
Conclusion: makeStableName is never called twice on the same value.
Hence it is useless and can be replaced by a simple id generator --
which we can do trivially because the graph construction algorithm
already runs in a state monad.

This removal of the usage of stable names is exactly what this commit
does. It fixes the corrupt graph issues for me.