Add annotation fields to the rest of OpenAcc
The parts here that we just handwaved away for now are:

- The sharing recovery
- Parts of fusion

Once everything else is in place we should check all uses of `Alet`,
`Avar`, and `mkDummyAnn`. `mkDummyAnn` should only very seldomly be
needed, and since it wouldn't be correct anyways all annotation fields
in the de Bruijn ASTs let and var constructors have been bruteforced
just to make the type checker happy.