Add annotation fields to the rest of OpenExp
Like in the previous commit doing the same thing for OpenAcc, there are
a couple of things that may not yet be correct and some things that most
definitely are broken. The sharing recovery and most things involving
let-bindings and `Evar` have been completely ignored and just use dummy
annotations at the moment so the type checker stops complaining.