Change and fix Elt sum type tag scheme
起
GHC Generics represents an ADT not as a list of lists of fields, but as
nested binary sums and products ((:+:) and (:*:)). This means that for
data types with more than 2 constructors or fields, an association must
be chosen. The association that GHC chooses in practice is an
approximately balanced binary tree, but the precise association is an
implementation detail.

承
Hence, the Generics machinery for Elt must be able to handle any
association. For products, the association was simply normalised to a
right-leaning tree with a CPS-like trick; this worked (and works) fine.
For sums, however, the same trick was used for the associativity of the
product in the representation (because Accelerate represents sums as
products), but the generated tags _were_ dependent on GHC's chosen
association. Specifically, in Haskell-ish pseudocode, the tag generation
followed this scheme:

  tags (a :+: b) = ((2*) <$> tags a) ++ ((\n -> 2*n+1) <$> tags b)
  tags _ = [0]

This generates unique, but association-dependent, tags for any binary
tree.

転
The pattern synonyms in Data.Array.Accelerate.Pattern(.TH) have to
generate, and match on, the tags generated by the Elt machinery, but
they live outside of the Elt class. The old code here was presumably
reverse-engineered from looking at what GHC generates for data types up
to 6 constructors, because it just so happens to match for data types of
6 constructors or less. It generates nonsense for larger data types.
It is hence no surprise that the reporter of issue #551 [1] noted that
data types of 6 constructors do work, but types of 7 constructors do
not.

Interesting to note might be that the reverse-engineered scheme in the
pattern synonym code resulted in exponential tag values, resulting in a
maximum number of constructors of around 16 (seeing as the tags are
stored in a Word8).

結
The Elt tag generation was refactored to extend the CPS-ish trick to tag
generation; the resulting scheme is "simply" assigning tags 0, 1, 2, ...
to the constructors of the data type in sequence. This allows data types
of up to 255 constructors, and makes the job of the pattern synonym code
easier too.

Naturally, this commit changes the representation of sum types, but
since those are essentially inaccessible from the public API of
accelerate and its backends (and any release of Accelerate that would
include this commit will have a new version number anyway, preventing
cache pollution), this is probably fine.

The only risk is that the new scheme makes too much sense and that
consumers will start to rely on it somehow. [2]

Fixes #551.

[1]: https://github.com/AccelerateHS/accelerate/issues/551#issuecomment-2783868398
[2]: https://xkcd.com/1172